Add files via upload
This python code will open an existed text file, created by tshark that captures data packets. The code will run and remove unnecessary text and write a file. Then creates one variable that looks for Ip address that occurs the most and another variable that displays the number of times it occurs from the text file.  Then if a most occuring ip address is greater than a limit, the raspberry pi will sound an alarm and notify the user.