Add files via upload
This project is a network surveillance system using Raspberry Pi as a monitoring node. The Raspberry Pi is configured to
perform packet capture using this python code. This code opens and reads a text file imported by tshark, that captures packets from Wifi.
The program searches for the IP address that appears the most from the text file and displays the number of times.
Additionally, the Raspberry Pi will sound an alarm and notify the user if there is an intrusion.